基于tf 2.x 环境的transformer模型可以跑起来了。 下一步：代码优化，逻辑重构

+++ tf_test.py
@@ -519,8 +519,8 @@ class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
     self.warmup_steps = warmup_steps
     
   def __call__(self, step):
-    arg1 = tf.math.rsqrt(step)
-    arg2 = step * (self.warmup_steps ** -1.5)
+    arg1 = tf.math.rsqrt(tf.cast(step, tf.float32))
+    arg2 = tf.cast(step,tf.float32) * (self.warmup_steps ** -1.5)
     
     return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
